+++ Frameworks/CoreFoundation.framework/__NSArrayReverseEnumerator.h
@@ -3,8 +3,8 @@
  */
 
 @interface __NSArrayReverseEnumerator : NSEnumerator {
-    unsigned int _idx;
-    id _obj;
+    unsigned long long  _idx;
+    id  _obj;
 }
 
 - (void)dealloc;
