+++ Frameworks/CoreFoundation.framework/__NSOrderedSetArrayProxy.h
@@ -3,13 +3,13 @@
  */
 
 @interface __NSOrderedSetArrayProxy : NSArray {
-    id _orderedSet;
+    id  _orderedSet;
 }
 
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
-- (unsigned int)count;
+- (unsigned long long)count;
 - (void)dealloc;
 - (id)initWithOrderedSet:(id)arg1;
-- (id)objectAtIndex:(unsigned int)arg1;
+- (id)objectAtIndex:(unsigned long long)arg1;
 
 @end
