+++ Frameworks/CoreFoundation.framework/__NSOrderedSetSetProxy.h
@@ -3,11 +3,11 @@
  */
 
 @interface __NSOrderedSetSetProxy : NSSet {
-    id _orderedSet;
+    id  _orderedSet;
 }
 
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
-- (unsigned int)count;
+- (unsigned long long)count;
 - (void)dealloc;
 - (id)initWithOrderedSet:(id)arg1;
 - (id)member:(id)arg1;
