+++ Frameworks/Foundation.framework/_NSNestedDictionary.h
@@ -3,13 +3,13 @@
  */
 
 @interface _NSNestedDictionary : NSMutableDictionary {
-    id _bindings;
-    id _locals;
+    id  _bindings;
+    id  _locals;
 }
 
 - (id)_recursiveAllKeys;
 - (id)_recursiveAllValues;
-- (unsigned int)count;
+- (unsigned long long)count;
 - (id)keyEnumerator;
 - (id)objectEnumerator;
 - (id)objectForKey:(id)arg1;
