+++ Frameworks/WebKit.framework/WKDateTimePopoverViewController.h
@@ -5,12 +5,12 @@
 @interface WKDateTimePopoverViewController : UIViewController {
     struct RetainPtr<NSObject<WKFormControl> > { 
         void *m_ptr; 
-    } _innerControl;
+    }  _innerControl;
 }
 
 - (id).cxx_construct;
 - (void).cxx_destruct;
-- (id)initWithView:(id)arg1 datePickerMode:(int)arg2;
+- (id)initWithView:(id)arg1 datePickerMode:(long long)arg2;
 - (id)innerControl;
 - (void)loadView;
