+++ Frameworks/WebKit.framework/WKSyntheticClickTapGestureRecognizer.h
@@ -3,15 +3,15 @@
  */
 
 @interface WKSyntheticClickTapGestureRecognizer : UITapGestureRecognizer {
-    SEL _gestureRecognizedAction;
-    id _gestureRecognizedTarget;
-    SEL _resetAction;
-    id _resetTarget;
+    SEL  _gestureRecognizedAction;
+    id  _gestureRecognizedTarget;
+    SEL  _resetAction;
+    id  _resetTarget;
 }
 
 - (void)reset;
 - (void)setGestureRecognizedTarget:(id)arg1 action:(SEL)arg2;
 - (void)setResetTarget:(id)arg1 action:(SEL)arg2;
-- (void)setState:(int)arg1;
+- (void)setState:(long long)arg1;
 
 @end
