+++ PrivateFrameworks/MobileBackup.framework/MBDigestSHA1.h
@@ -13,12 +13,12 @@
         unsigned int Nh; 
         unsigned int data[16]; 
         int num; 
-    } _context;
+    }  _context;
 }
 
 - (id)final;
-- (void)finalWithBytes:(void*)arg1 length:(unsigned int)arg2;
+- (void)finalWithBytes:(void*)arg1 length:(unsigned long long)arg2;
 - (id)init;
-- (void)updateWithBytes:(const void*)arg1 length:(unsigned int)arg2;
+- (void)updateWithBytes:(const void*)arg1 length:(unsigned long long)arg2;
 
 @end
