+++ PrivateFrameworks/OfficeImport.framework/ECIntTwoIntKeyDictionary.h
@@ -3,10 +3,10 @@
  */
 
 @interface ECIntTwoIntKeyDictionary : OITSUNoCopyDictionary {
-    id mNilKeyObject;
+    id  mNilKeyObject;
 }
 
-- (id)initWithCapacity:(unsigned int)arg1;
+- (id)initWithCapacity:(unsigned long long)arg1;
 - (id)objectForKey:(id)arg1;
 - (void)setObject:(id)arg1 forKey:(id)arg2;
 - (void)setObject:(id)arg1 forUncopiedKey:(id)arg2;
