+++ PrivateFrameworks/OfficeImport.framework/OISFUZipFreeSpaceEntry.h
@@ -3,10 +3,10 @@
  */
 
 @interface OISFUZipFreeSpaceEntry : NSObject {
-    unsigned long long length;
-    unsigned long long offset;
+    unsigned long long  length;
+    unsigned long long  offset;
 }
 
-- (int)compareByOffset:(id)arg1;
+- (long long)compareByOffset:(id)arg1;
 
 @end
